feat: move check methods to disdrodb.api.checks
